Circuit Image - See notes!
FYI Power is pretty important here - I used a 3.7 V battery in case that isnt visible.  As well, fritzing didnt have the neopixel matrix I used.  It was an 8x8 from Adafruit, and there are three inputs, power, ground, and data in.  Then there is another set of 3 that has a data out label - that is for chaining the matrices together.  I didnt use that.  I attached pin D2 to Data in, I didnt mention this on the drawing because the matrix was not the same.